@@ -1,18 +1,20 @@
 {
-  "name": "my-app",
+  "name": "witcher",
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "@ramonak/react-progress-bar": "^4.0.0",
-    "@testing-library/jest-dom": "^5.14.1",
-    "@testing-library/react": "^11.2.7",
-    "@testing-library/user-event": "^12.8.3",
-    "classnames": "^2.3.1",
+    "@pbe/react-yandex-maps": "^1.2.2",
+    "@testing-library/jest-dom": "^5.16.5",
+    "@testing-library/react": "^13.4.0",
+    "@testing-library/user-event": "^13.5.0",
     "formik": "^2.2.9",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
-    "react-scripts": "4.0.3",
-    "web-vitals": "^1.1.2",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-loader-spinner": "^5.3.4",
+    "react-scripts": "5.0.1",
+    "react-select": "^5.7.0",
+    "swiper": "^8.4.6",
+    "web-vitals": "^2.1.4",
     "yup": "^0.32.11"
   },
   "scripts": {
@@ -38,5 +40,8 @@
       "last 1 firefox version",
       "last 1 safari version"
     ]
+  },
+  "devDependencies": {
+    "react-router-dom": "^6.7.0"
   }
 }
