@@ -1,135 +1,162 @@
-import React, { useRef } from "react";
-import paperclip from "../../images/paperclip.png";
+import paperclip from "../../assets/images/paperclip.png";
 import { Formik, Form, Field, ErrorMessage } from "formik";
 import * as Yup from "yup";
+import { TextError, InputFormik } from "./InputFormik";
+import { useState } from "react";
+import CitySelect from "./CitySelect";
 
-import { RequestFormikControl, TextError } from "./RequestFormikControl";
-import { NavLink } from "react-router-dom";
-
-const phoneRegExp =
-  /^((\\+[1-9]{1,4}[ \\-]*)|(\\([0-9]{2,3}\\)[ \\-]*)|([0-9]{2,4})[ \\-]*)*?[0-9]{3,4}?[ \\-]*[0-9]{3,4}?$/;
 
 const validationSchema = Yup.object().shape({
-  city: Yup.string().required("Город не выбран"),
+  city: Yup.object()
+    .shape({
+      value: Yup.string(),
+      label: Yup.string(),
+    })
+    .required("Поле не заполнено")
+    .nullable(),
   name: Yup.string().required("Поле не заполнено"),
   remark: Yup.string().required("Поле не заполнено"),
   email: Yup.string()
-    .email("Email должен содержать @")
+    .email("Email должен содержать @ и имя домена")
     .required("Поле не заполнено"),
+  number: Yup.number().required("Поле не заполнено"),
+  file: Yup.array().min(1, "Поле не заполнено"),
   toggle: Yup.array().min(1, "Поле не заполнено"),
-  number: Yup.string()
-    .required("Поле не заполнено")
-    .matches(phoneRegExp, "Неверный формат")
-    .min(5, "Слишком короткий номер")
-    .max(15, "Слишком длинный номер"),
 });
 
-export default function RequestFormikContainer(props) {
-  const cityOptions = [
-    { key: "Выберите город", value: "" },
-    { key: "Москва", value: "Москва" },
-    { key: "Санкт-Петербург", value: "Санкт-Петербург" },
-    { key: "Казань", value: "Казань" },
-    { key: "Краснодар", value: "Краснодар" },
-    { key: "Ростов-на-Дону", value: "Ростов-на-Дону" },
-  ];
 
+export default function RequestFormikContainer(props) {
+  const [labelValue, setLabelValue] = useState(["Прикрепите файл (до 3 шт.)"]);
   return (
     <div>
-      <div className="request_title">
-        <h1>Оставьте заявку</h1>
-      </div>
+      <h1>Оставьте заявку</h1>
       <div>
         <Formik
           enableReinitialize={true}
           initialValues={{
-            city: "",
+            city: null,
             name: "",
             email: "",
             number: "",
             remark: "",
-            file: "",
+            file: [],
             toggle: [],
           }}
           validationSchema={validationSchema}
-          onSubmit={(values, onSubmitProps) => {
-            onSubmitProps.setSubmitting(false);
-            console.log(values);
+          onSubmit={(values, formikBag) => {
+            console.log("values", values);
+            formikBag.setTouched(true);
+            formikBag.setSubmitting(false);
+            props.goToRequestIsSent();
           }}
         >
-          {({ errors, touched, values, setFieldValue, isValid }) => {
+          {({ errors, touched, values, setFieldValue, setTouched }) => {
             return (
               <Form>
-                <div className="form_input_wrapper">
-                  <RequestFormikControl
-                    control="select"
-                    type="select"
-                    label="city"
+                <div className="select_wrapper">
+                  <CitySelect
                     name="city"
-                    options={cityOptions}
-                    className="form_input"
-                    errors={errors}
+                    error={errors.city}
                     touched={touched}
+                    value={values.city}
+                    setFieldValue={setFieldValue}
+                    setTouched={setTouched}
                   />
+                  <ErrorMessage name={"city"} component={TextError} />
                 </div>
-
                 <div className="form_input_wrapper">
-                  <RequestFormikControl
-                    control="input"
+                  <InputFormik
                     type="name"
                     label="name"
                     name="name"
                     placeholder={"Имя"}
-                    className="form_input"
+                    className={
+                      touched.name && errors.name
+                        ? "form_input invalid"
+                        : "form_input valid"
+                    }
                   />
                 </div>
                 <div className="form_input_inline_group">
-                  {/* <div > */}
-                  <RequestFormikControl
-                    control="input"
-                    type="email"
-                    label="email"
-                    name="email"
-                    placeholder={"Email"}
-                    className="form_input_half"
-                  />
-                  <RequestFormikControl
-                    control="input"
-                    label="number"
-                    name="number"
-                    placeholder={"+7(___)__-__-___"}
-                    className="form_input_half"
-                  />
+                  <div>
+                    <InputFormik
+                      type="email"
+                      label="email"
+                      name="email"
+                      placeholder={"Email"}
+                      className={
+                        touched.email && errors.email
+                          ? "form_input invalid"
+                          : "form_input valid"
+                      }
+                    />
+                  </div>
+                  <div>
+                    <InputFormik
+                      label="number"
+                      name="number"
+                      placeholder={"+7(___)__-__-___"}
+                      className={
+                        touched.number && errors.number
+                          ? "form_input invalid"
+                          : "form_input valid"
+                      }
+                    />
+                  </div>
                 </div>
                 <div className="form_textArea_wrapper">
-                  <Field
-                    className="form_textArea"
-                    name="remark"
+                  <InputFormik
                     label="remark"
+                    name="remark"
                     placeholder={"Оставьте пометку к заказу"}
+                    errors={errors}
+                    touched={touched}
+                    className={
+                      touched.remark && errors.remark
+                        ? "form_textArea invalid"
+                        : "form_textArea valid"
+                    }
                   />
-                  <ErrorMessage name="remark" component={TextError} />
                 </div>
-
-                <div>
-                  <div className="form_file">
-                    <label for="file">
-                      {values.file.name ? (
-                        <p>{values.file.name}</p>
-                      ) : (
-                        <p>Прикрепите файл</p>
-                      )}
-                    </label>
-                    <label class="custom_file_upload">
+                <div className="form_file_wrapper">
+                  <div
+                    className={
+                      touched.file && errors.file
+                        ? "form_file invalid"
+                        : "form_file valid"
+                    }
+                    tabIndex={0}
+                    onBlur={(e) => {
+                      setTouched({ ...touched, file: true });
+                    }}
+                  >
+                    <label htmlFor="file">
+                      {labelValue.map((elem) => (
+                        <span key={elem}>{elem}</span>
+                      ))}
                       <input
                         id="file"
                         name="file"
                         type="file"
                         onChange={(event) => {
-                          setFieldValue("file", event.currentTarget.files[0]);
+                          let file = event.target.files[0];
+                          if (file) {
+                            setLabelValue(
+                              [
+                                ...labelValue.filter(
+                                  (item) =>
+                                    item !== "Прикрепите файл (до 3 шт.)"
+                                ),
+                                file.name,
+                              ].slice(0, 3)
+                            );
+                            let currentFilesCopy = values.file.slice(0); // to get a copy of the original array
+                            currentFilesCopy.push(file);
+                            setFieldValue("file", currentFilesCopy);
+                            setTouched({ ...touched, file: true });
+                          }
                         }}
                       />
-
                       <img
                         src={paperclip}
                         alt="paperclip"
@@ -137,37 +164,32 @@ export default function RequestFormikContainer(props) {
                       />
                     </label>
                   </div>
-                  <div>{touched.file ? <p>hyju</p> : null}</div>
-
                   <ErrorMessage name="file" component={TextError} />
-
-                  <div className="checkbox_wrapper">
-                    <label class="checkbox_container">
-                      <Field type="checkbox" name="toggle" value="Checked" />
-                      <p>
-                        {"Даю согласие на обработку своих персональных данных"}
-                      </p>
-                      <span class="checkmark"></span>
-                    </label>
-                    <ErrorMessage name="toggle" component={TextError} />
-                  </div>
-
-                  <div>
-                    <button
-                      type="submit"
-                      onClick={props.goToRequestIsSent}
-                      disabled={
-                        Object.keys(touched).length !== 0 ? !isValid : "true"
-                      }
+                </div>
+                <div className="checkbox_wrapper">
+                  <label className="checkbox_container">
+                    <Field type="checkbox" name="toggle" value="Checked" />
+                    <p>
+                      {"Даю согласие на обработку своих персональных данных"}
+                    </p>
+                    <span
                       className={
-                        Object.keys(touched).length === 0 || isValid === false
-                          ? "button_dis"
-                          : "button"
+                        touched.toggle && errors.toggle
+                          ? "checkmark invalid"
+                          : "checkmark valid"
                       }
-                    >
-                      Оставить заявку
-                    </button>
-                  </div>
+                      tabIndex={0}
+                      onBlur={(e) => {
+                        setTouched({ ...touched, toggle: true });
+                      }}
+                    ></span>
+                  </label>
+                  <ErrorMessage name="toggle" component={TextError} />
+                </div>
+                <div>
+                  <button type="submit" className={"button_default_red"}>
+                    Оставить заявку
+                  </button>
                 </div>
               </Form>
             );
